modified spark icon, reset password and removed post related parts to components
