+++ assets/sparkle.json
@@ -1,0 +1,1 @@
+{"v":"4.8.0","meta":{"g":"LottieFiles AE ","a":"","k":"","d":"","tc":""},"fr":24,"ip":0,"op":48,"w":600,"h":600,"nm":"TWINKLE","ddd":0,"assets":[],"layers":[{"ddd":0,"ind":1,"ty":4,"nm":"Shape Layer 10","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[300,300,0],"ix":2},"a":{"a":0,"k":[0,0,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.297,0.297,0.667],"y":[1.046,1.046,1]},"o":{"x":[0.167,0.167,0.167],"y":[0.167,0.167,17.014]},"t":0,"s":[0,0,100]},{"i":{"x":[0.833,0.833,0.833],"y":[0.773,0.773,24.188]},"o":{"x":[0.226,0.226,0.333],"y":[-0.013,-0.013,0]},"t":7.955,"s":[49,49,100]},{"t":38,"s":[0,0,100]}],"ix":6}},"ao":0,"shapes":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"v":[[0,-374],[10.566,-25.509],[99.702,-89.936],[25.509,-10.566],[374,0],[25.509,10.566],[114.26,106.26],[10.566,25.509],[0,374],[-10.566,25.509],[-87.233,80.675],[-25.509,10.566],[-374,0],[-25.509,-10.566],[-91.558,-79.702],[-10.566,-25.509]],"c":true},"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"pb","nm":"Pucker & Bloat 1","a":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":0,"s":[10]},{"t":38,"s":[-111]}],"ix":1},"ix":2,"mn":"ADBE Vector Filter - PB","hd":false},{"ty":"fl","c":{"a":0,"k":[1,1,1,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false}],"ip":0,"op":43,"st":0,"bm":0},{"ddd":0,"ind":2,"ty":4,"nm":"Shape Layer 12","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[300,300,0],"ix":2},"a":{"a":0,"k":[0,0,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.297,0.297,0.667],"y":[1.046,1.046,1]},"o":{"x":[0.167,0.167,0.167],"y":[0.167,0.167,17.014]},"t":1,"s":[0,0,100]},{"i":{"x":[0.833,0.833,0.833],"y":[0.78,0.78,23.416]},"o":{"x":[0.226,0.226,0.333],"y":[-0.013,-0.013,0]},"t":8.955,"s":[49,49,100]},{"t":38,"s":[0,0,100]}],"ix":6}},"ao":0,"ef":[{"ty":29,"nm":"Gaussian Blur","np":5,"mn":"ADBE Gaussian Blur 2","ix":1,"en":1,"ef":[{"ty":0,"nm":"Blurriness","mn":"ADBE Gaussian Blur 2-0001","ix":1,"v":{"a":0,"k":78.3,"ix":1}},{"ty":7,"nm":"Blur Dimensions","mn":"ADBE Gaussian Blur 2-0002","ix":2,"v":{"a":0,"k":1,"ix":2}},{"ty":7,"nm":"Repeat Edge Pixels","mn":"ADBE Gaussian Blur 2-0003","ix":3,"v":{"a":0,"k":0,"ix":3}}]}],"shapes":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"v":[[0,-374],[10.566,-25.509],[99.702,-89.936],[25.509,-10.566],[374,0],[25.509,10.566],[114.26,106.26],[10.566,25.509],[0,374],[-10.566,25.509],[-87.233,80.675],[-25.509,10.566],[-374,0],[-25.509,-10.566],[-91.558,-79.702],[-10.566,-25.509]],"c":true},"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"pb","nm":"Pucker & Bloat 1","a":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":1,"s":[10]},{"t":38,"s":[-111]}],"ix":1},"ix":2,"mn":"ADBE Vector Filter - PB","hd":false},{"ty":"fl","c":{"a":0,"k":[0.9764705882352941,0.7254901960784313,1,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false}],"ip":1,"op":44,"st":1,"bm":0},{"ddd":0,"ind":3,"ty":4,"nm":"Shape Layer 11","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[300,300,0],"ix":2},"a":{"a":0,"k":[0,0,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.297,0.297,0.667],"y":[1.046,1.046,1]},"o":{"x":[0.167,0.167,0.167],"y":[0.167,0.167,17.014]},"t":2,"s":[0,0,100]},{"i":{"x":[0.833,0.833,0.833],"y":[0.788,0.788,22.645]},"o":{"x":[0.226,0.226,0.333],"y":[-0.012,-0.012,0]},"t":9.955,"s":[49,49,100]},{"t":38,"s":[0,0,100]}],"ix":6}},"ao":0,"ef":[{"ty":29,"nm":"Gaussian Blur","np":5,"mn":"ADBE Gaussian Blur 2","ix":1,"en":1,"ef":[{"ty":0,"nm":"Blurriness","mn":"ADBE Gaussian Blur 2-0001","ix":1,"v":{"a":0,"k":78.3,"ix":1}},{"ty":7,"nm":"Blur Dimensions","mn":"ADBE Gaussian Blur 2-0002","ix":2,"v":{"a":0,"k":1,"ix":2}},{"ty":7,"nm":"Repeat Edge Pixels","mn":"ADBE Gaussian Blur 2-0003","ix":3,"v":{"a":0,"k":0,"ix":3}}]}],"shapes":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"v":[[0,-374],[10.566,-25.509],[99.702,-89.936],[25.509,-10.566],[374,0],[25.509,10.566],[114.26,106.26],[10.566,25.509],[0,374],[-10.566,25.509],[-87.233,80.675],[-25.509,10.566],[-374,0],[-25.509,-10.566],[-91.558,-79.702],[-10.566,-25.509]],"c":true},"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"pb","nm":"Pucker & Bloat 1","a":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":2,"s":[10]},{"t":38,"s":[-111]}],"ix":1},"ix":2,"mn":"ADBE Vector Filter - PB","hd":false},{"ty":"fl","c":{"a":0,"k":[0.8666666666666667,0.3411764705882353,1,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false}],"ip":2,"op":45,"st":2,"bm":0},{"ddd":0,"ind":4,"ty":4,"nm":"Shape Layer 9","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[300,300,0],"ix":2},"a":{"a":0,"k":[0,0,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.297,0.297,0.667],"y":[1.046,1.046,1]},"o":{"x":[0.167,0.167,0.167],"y":[0.167,0.167,17.014]},"t":3,"s":[0,0,100]},{"i":{"x":[0.833,0.833,0.833],"y":[0.796,0.796,21.873]},"o":{"x":[0.226,0.226,0.333],"y":[-0.012,-0.012,0]},"t":10.955,"s":[49,49,100]},{"t":38,"s":[0,0,100]}],"ix":6}},"ao":0,"ef":[{"ty":29,"nm":"Gaussian Blur","np":5,"mn":"ADBE Gaussian Blur 2","ix":1,"en":1,"ef":[{"ty":0,"nm":"Blurriness","mn":"ADBE Gaussian Blur 2-0001","ix":1,"v":{"a":0,"k":78.3,"ix":1}},{"ty":7,"nm":"Blur Dimensions","mn":"ADBE Gaussian Blur 2-0002","ix":2,"v":{"a":0,"k":1,"ix":2}},{"ty":7,"nm":"Repeat Edge Pixels","mn":"ADBE Gaussian Blur 2-0003","ix":3,"v":{"a":0,"k":0,"ix":3}}]}],"shapes":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0],[0,0]],"v":[[0,-374],[10.566,-25.509],[99.702,-89.936],[25.509,-10.566],[374,0],[25.509,10.566],[114.26,106.26],[10.566,25.509],[0,374],[-10.566,25.509],[-87.233,80.675],[-25.509,10.566],[-374,0],[-25.509,-10.566],[-91.558,-79.702],[-10.566,-25.509]],"c":true},"ix":2},"nm":"Path 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"pb","nm":"Pucker & Bloat 1","a":{"a":1,"k":[{"i":{"x":[0.833],"y":[0.833]},"o":{"x":[0.167],"y":[0.167]},"t":3,"s":[10]},{"t":38,"s":[-111]}],"ix":1},"ix":2,"mn":"ADBE Vector Filter - PB","hd":false},{"ty":"fl","c":{"a":0,"k":[1,1,1,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false}],"ip":3,"op":46,"st":3,"bm":0}],"markers":[]}
