+++ package.json
@@ -23,6 +23,7 @@
     "expo-notifications": "^0.29.8",
     "expo-status-bar": "~1.11.1",
     "firebase": "^11.0.1",
+    "lottie-react-native": "7.1.0",
     "react": "18.2.0",
     "react-native": "0.76.2",
     "react-native-deck-swiper": "^2.0.17",
