+++ src/styles/PostStyles.js
@@ -191,53 +191,51 @@ const postScreenStyles = StyleSheet.create({
 // Styles for CreatePostScreen
 const createPostScreenStyles = StyleSheet.create({
     createPostContainer: {
-        flex: 1,
-        backgroundColor: '#fff',
-    },
-    createPostContainerHeader: {
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        paddingHorizontal: 15,
-        paddingVertical: 10,
-        borderBottomWidth: 1,
-        borderBottomColor: '#f0f0f0',
+      flex: 1,
+      backgroundColor: '#fff',
     },
     createPostContent: {
-        flex: 1,
-        padding: 15,
+      flex: 1,
+      padding: 15,
+      marginBottom: 70, 
     },
     createPostInput: {
-        fontSize: 16,
-        color: '#333',
-        minHeight: 100,
+      fontSize: 16,
+      color: '#333',
+      minHeight: 100,
     },
     createImagePreviewContainer: {
-        marginTop: 15,
-        position: 'relative',
+      marginTop: 15,
+      position: 'relative',
     },
     createImagePreview: {
-        width: '100%',
-        height: 300,
-        borderRadius: 10,
+      width: '100%',
+      height: 300,
+      borderRadius: 10,
     },
     createPostFooter: {
-        flexDirection: 'row',
-        padding: 15,
-        borderTopWidth: 1,
-        borderTopColor: '#f0f0f0',
+      flexDirection: 'row',
+      justifyContent: 'space-around', 
+      alignItems: 'center',
+      paddingVertical: 10,
+      borderTopWidth: 1,
+      borderTopColor: '#f0f0f0',
+      backgroundColor: '#fff',
     },
     createPostMediaButton: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        marginRight: 20,
-        padding: 5,
+      flexDirection: 'column',
+      alignItems: 'center',
+      justifyContent: 'center',
+      paddingVertical: 8,
+      width: 60,
     },
     createPostMediaButtonText: {
-        marginLeft: 5,
-        color: '#FF69B4',
-        fontWeight: '500',
+      marginTop: 4,
+      color: '#FF69B4',
+      fontSize: 12,
+      fontWeight: '500',
+      textAlign: 'center',
     },
-});
+  });
 
 export { createPostScreenStyles, postScreenStyles };
