+++ src/utils/timeFormatter.js
@@ -1,0 +1,15 @@
+export const formatTime = (timestamp) => {
+    if (!timestamp || !timestamp.toDate) return '';
+    
+    const now = new Date();
+    const postTime = timestamp.toDate();
+    const diffMs = now - postTime;
+    const diffMins = Math.floor(diffMs / (1000 * 60));
+    const diffHours = Math.floor(diffMins / 60);
+    const diffDays = Math.floor(diffHours / 24);
+  
+    if (diffMins < 1) return 'Just now';
+    if (diffMins < 60) return `${diffMins} min ago`;
+    if (diffHours < 24) return `${diffHours} hours ago`;
+    return `${diffDays} days ago`;
+};
